Get instant serach filters to play nice with Modals
