+++ site/gatsby-site/src/components/discover/Filters.js
@@ -1,37 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import REFINEMENT_LISTS from 'components/discover/REFINEMENT_LISTS';
-import { Row, Col, Modal, Button } from 'react-bootstrap';
-import styled from 'styled-components';
-import RangeInput from './filters/RangeInput';
-import RefinementList from './filters/RefinementList';
-
-function Filter({ attribute, ...rest }) {
-  let Component = null;
-
-  switch (attribute) {
-    case 'epoch_incident_date':
-    case 'epoch_date_published':
-      Component = RangeInput;
-      break;
-    default:
-      Component = RefinementList;
-      break;
-  }
-
-  return <Component attribute={attribute} {...rest} />;
-}
-
-const FiltersModalList = styled.div`
-  gap: 0.5rem;
-  display: flex;
-  flex-direction: column;
-`;
+import { Row, Col } from 'react-bootstrap';
+import Filter from './Filter';
 
 function Filters() {
-  const [showModal, setShowModal] = useState(false);
-
-  const handleClose = () => setShowModal(false);
-
   return (
     <>
       <Row className="d-none d-md-flex justify-content-md-between mb-3">
@@ -39,7 +11,7 @@ function Filters() {
           {['classifications', 'source_domain', 'authors', 'submitters', 'incident_id']
             .map((a) => REFINEMENT_LISTS.find((list) => list.attribute == a))
             .map((list) => (
-              <Filter key={list.attribute} attribute={list.attribute} {...list} />
+              <Filter key={list.attribute} type={list.type} {...list} />
             ))}
         </Col>
         <Col className="d-flex gap-1 mt-2 col-md-auto">
@@ -50,32 +22,6 @@ function Filters() {
             ))}
         </Col>
       </Row>
-
-      <Row className="my-3 d-md-none">
-        <Col className="d-flex justify-content-end">
-          <Button variant="link" onClick={() => setShowModal(true)}>
-            Filters
-          </Button>
-        </Col>
-      </Row>
-
-      <Modal show={showModal} onHide={handleClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>Filters</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <FiltersModalList>
-            {REFINEMENT_LISTS.map((list) => (
-              <Filter key={list.attribute} attribute={list.attribute} {...list} />
-            ))}
-          </FiltersModalList>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Close
-          </Button>
-        </Modal.Footer>
-      </Modal>
     </>
   );
 }
